refactor(user): 유저 enums 패키지 분리
